More renames for clarity
