+++ src/actions/ConnectionHandler.ts
@@ -9,7 +9,7 @@
  */
 
 import { ConnectionStore } from "../storage/ConnectionStore";
-import { connectFTPServer } from "../services/utils";
+import { connectFTPServer } from "../services/ServiceUtils";
 import { IIpcConnectionArgs, IResponse } from '../types/interfaces';
 
 class Connection {
