+++ src/renderer/components/Home.tsx
@@ -20,7 +20,7 @@ import { Tooltip } from '@mui/material';
 import ArrowCircleRightIcon from '@mui/icons-material/ArrowCircleRight';
 import installationImg from '../assets/installation.png'
 import installationDryImg from '../assets/installation-dry-run.png'
-import eventDispatcher from "../../utils/eventDispatcher";
+import eventDispatcher from "../../services/eventDispatcher";
 import { selectActiveStepIndex, selectIsSubstep, selectActiveSubStepIndex} from './stages/progress/activeStepSlice';
 import { selectConnectionStatus} from './stages/progress/progressSlice';
 import { selectActiveStepDate} from './stages/progress/activeStepSlice';
