+++ src/services/CheckJava.ts
@@ -8,9 +8,10 @@
  * Copyright Contributors to the Zowe Project.
  */
 
+import { JCL_UNIX_SCRIPT_OK } from "../renderer/components/common/Utils";
 import {IIpcConnectionArgs, IJobResults} from "../types/interfaces";
 import {submitJcl} from "./SubmitJcl";
-import {JCL_UNIX_SCRIPT_OK, startBPXBATCHAndShellSession} from "./utils";
+import { startBPXBATCHAndShellSession } from "./ServiceUtils";
 
 export class CheckJava {
