+++ src/services/ServiceUtils.ts
@@ -13,12 +13,9 @@ import Header from "../renderer/components/Header"
 import { AlertColor } from "@mui/material/Alert";
 import zos from 'zos-node-accessor';
 
-// (GLOBAL NOTE): Due to the structure of React, there may be some needed duplication with zen\src\renderer\components\common\Utils.tsx !
+// Note: This file is not usable by the Renderer
 
 export const JCL_UNIX_SCRIPT_CHARS = 70;
-
-// Note: This doesn't gaurantee command did what it was supposed to do, but rather z/OS Unix (and zwe) didn't throw an error
-export const JCL_UNIX_SCRIPT_OK = "Script finished.";
 
 export const JCL_JOBNAME_DEFAULT = "ZENJOB";
